extract_data_from_sources/requirements.txt
pandas
requests
astropy
dustmaps
python-dotenv